html/cross-origin-opener-policy/js-url.html

<!doctype html>
<title>Cross-Origin-Opener-Policy and a JavaScript URL popup</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/get-host-info.sub.js></script>
<script src="/common/utils.js"></script>
<script>
function setup(t, coop, name) {
  window.furtherPopup = null;

  const bc = new BroadcastChannel(name);

  const popup = window.open(`javascript:\`<script>
const w = window.open("${get_host_info().HTTPS_REMOTE_ORIGIN}/html/cross-origin-opener-policy/resources/coop-coep.py?coop=x&coep=x&channel=${bc.name}", "${bc.name}");
window.opener.furtherPopup = w;
<\/script>\``);
  t.add_cleanup(() => {
    // Close the popups once the test is complete.
    // The browsing context of the second popup is closed hence use the
    //  broadcast channel to trigger the closure.
    bc.postMessage('close');
    popup.close();
  });

  return Promise.all([
    new Promise((resolve) => popup.onload = () => resolve(popup)),
    new Promise((resolve) => bc.onmessage = ({data}) => resolve(data))
  ]);

  return Promise.all([whenLoaded, whenMessage]);
}
promise_test(t => {
  const name = token();
  return setup(t, 'x', name)
    .then(([popup, data]) => {
      assert_equals(popup.opener, window);
      assert_equals(popup.origin, window.origin);
      assert_equals(data.name, '');
      assert_false(data.opener);
      assert_true(furtherPopup.closed);
      assert_equals(furtherPopup.document.URL, 'about:blank');
    });
}, 'coop: x');

promise_test(t => {
  const name = token();
  return setup(t, 'same-origin', name)
    .then(([popup, data]) => {
      assert_equals(popup.opener, window);
      assert_equals(popup.origin, window.origin);
      assert_equals(data.name, name);
      assert_true(data.opener);
      assert_true(furtherPopup.closed);
      assert_equals(furtherPopup.document.URL, 'about:blank');
    });
}, 'coop: same-origin');
</script>
